+++ src/Dbflute/BsEntity/BsCustomer.cs
@@ -55,8 +55,8 @@ namespace Dbflute.ExEntity {
         //                                                                       Attribute
         //                                                                       =========
         #region Attribute
-        /// <summary>id: {PK, ID, NotNull, serial(10)}</summary>
-        protected int? _id;
+        /// <summary>id: {PK, ID, NotNull, bigserial(19)}</summary>
+        protected long? _id;
 
         /// <summary>last_name: {NotNull, varchar(20)}</summary>
         protected String _lastName;
@@ -173,10 +173,10 @@ namespace Dbflute.ExEntity {
         //                                                                        Accessor
         //                                                                        ========
         #region Accessor
-        /// <summary>id: {PK, ID, NotNull, serial(10)}</summary>
+        /// <summary>id: {PK, ID, NotNull, bigserial(19)}</summary>
         [Seasar.Dao.Attrs.ID("identity")]
         [Seasar.Dao.Attrs.Column("id")]
-        public int? Id {
+        public long? Id {
             get { return _id; }
             set {
                 __modifiedProperties.AddPropertyName("Id");
